+++ lib/rley/ptree/terminal_node.rb
@@ -8,7 +8,10 @@ module Rley # This module is used as a namespace
 
       # aPosition is the position of the token in the input stream.      
       def initialize(aToken, aPos)
-        range = aPos.kind_of?(Fixnum) ? { low: aPos, high: aPos + 1 } : aPos
+        (major, minor) =  
+        
+        # Use '1.class' trick to support both Integer and Fixnum classes
+        range = aPos.kind_of?(1.class) ? { low: aPos, high: aPos + 1 } : aPos
         super(aToken.terminal, range)
         @token = aToken
       end